patch delete task invalid form submit.

+++ client/src/views/DeleteTask.vue
@@ -61,15 +61,18 @@ export default Vue.extend({
     };
   },
   computed: {
-    taskIdState(): boolean {
+    taskIdState: function (): boolean {
       return /^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i.test(
         this.taskId
       );
     }
   },
   methods: {
-    async onSubmit(event: any): Promise<void> {
+    onSubmit: async function (event: any): Promise<void> {
       event.preventDefault();
+      if (!this.taskIdState) {
+        return;
+      }
       const id: string = await taskService.deleteTask(this.taskId);
       if (id == this.taskId) {
         console.log(`Deleted task with id: ${id}`);
